+++ LibCarla/source/carla/trafficmanager/TrafficManagerLocal.cpp
@@ -37,6 +37,7 @@ TrafficManagerLocal::TrafficManagerLocal(
                                          track_traffic,
                                          local_map,
                                          parameters,
+                                         marked_for_removal,
                                          localization_frame,
                                          debug_helper)),
 
@@ -71,6 +72,7 @@ TrafficManagerLocal::TrafficManagerLocal(
     alsm(ALSM(registered_vehicles,
               buffer_map,
               track_traffic,
+              marked_for_removal,
               parameters,
               world,
               local_map,
@@ -123,7 +125,7 @@ void TrafficManagerLocal::Run() {
     // Wait for external trigger to initiate cycle in synchronous mode.
     if (synchronous_mode) {
       std::unique_lock<std::mutex> lock(step_execution_mutex);
-      step_begin_trigger.wait(lock, [this]() {return step_begin.load();});
+      step_begin_trigger.wait(lock, [this]() {return step_begin.load() || !run_traffic_manger.load();});
       step_begin.store(false);
     }
 
@@ -218,6 +220,9 @@ bool TrafficManagerLocal::SynchronousTick() {
 void TrafficManagerLocal::Stop() {
 
   run_traffic_manger.store(false);
+  if (parameters.GetSynchronousMode()) {
+    step_begin_trigger.notify_one();
+  }
 
   if (worker_thread) {
     if (worker_thread->joinable()) {
@@ -335,6 +340,10 @@ void TrafficManagerLocal::SetHybridPhysicsRadius(const float radius) {
   parameters.SetHybridPhysicsRadius(radius);
 }
 
+void TrafficManagerLocal::SetOSMMode(const bool mode_switch) {
+  parameters.SetOSMMode(mode_switch);
+}
+
 bool TrafficManagerLocal::CheckAllFrozen(TLGroup tl_to_freeze) {
   for (auto &elem : tl_to_freeze) {
     if (!elem->IsFrozen() || elem->GetState() != TLS::Red) {
@@ -382,7 +391,12 @@ void TrafficManagerLocal::ResetAllTrafficLights() {
 }
 
 void TrafficManagerLocal::SetSynchronousMode(bool mode) {
+  const bool previous_mode = parameters.GetSynchronousMode();
   parameters.SetSynchronousMode(mode);
+  if (previous_mode && !mode) {
+    step_begin.store(true);
+    step_begin_trigger.notify_one();
+  }
 }
 
 void TrafficManagerLocal::SetSynchronousModeTimeOutInMiliSecond(double time) {