+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 ## CARLA 0.9.10
 
-
+  * Support for OSM in Traffic Manager
+  * Fixed a map change error when Traffic Manager is in synchronous mode
+  * Fixes add entry issue for applying parameters more than once in Traffic Manager
+  * Fixes std::numeric_limits<float>::epsilon error in Traffic Manager
   * Upgraded carla Docker image to Ubuntu 18.04
   * Fixed memory leak on manual_control scripts (sensor listenning was not stoped before destroying)
   * Added PythonAPI `carla.Osm2Odr.convert()` function and `calra.Osm2OdrSettings` class to support Open Street Maps to OpenDRIVE conversions
@@ -19,12 +22,15 @@
   * Fixed rain drop spawn issues when spawning camera sensors
   * Fixed assets import pipeline
   * Fixed Update.sh from failing when the root folder contains a space on it
+  * Fixed dynamic meshes not moving to initial position when replaying
   * Fixed colors of lane markings when importing a map, they were reversed (white and yellow)
   * Fixed missing include directive in file **WheelPhysicsControl.h**
   * Fixed gravity measurement bug from IMU sensor
   * All sensors are now multi-stream, that means that the same sensor can be listened from different clients
   * Fixed point cloud of LiDAR. Now the points are given correctly in the sensor's coordinate system
   * Fixed light intensity and camera parameters to match
+  * Fixed and improved auto exposure camera (`histogram` exposure mode)
+  * RGB cameras `exposure_mode` is now set to `histogram` by default
   * New Lidar sensor ('lidar.ray_cast_semantic') that returns the point cloud with information regarding to the object that have collided: incident angle, idx of collided actor and it semantic tag
   * Added `opend3D.py`, a more friendly LiDAR visualizer
   * Exposed matrix form of transformation to the client and Python API
@@ -52,7 +58,9 @@
   * Replace deprectated `platform.dist()` with recommended `distro.linux_distribution()`
   * Improved the performance on capture sensors.
   * Fixed minor typo in the introduction section of documentation.
+  * Fixed a bug at the local planner when changing the route, causing it to maintain the first part of the previous one. This was only relevant when using very large buffer sizes.
   * Added automatic calculation of vehicle's BB
+
 
 ## CARLA 0.9.9